+++ src/Symfony/Component/Messenger/CHANGELOG.md
@@ -19,6 +19,8 @@ CHANGELOG
  * Add the `messenger:show` command to list and inspect pending messages of a transport
  * Make the `messenger:consume` and `messenger:failed:retry` commands exit immediately when a second `SIGINT` is received
  * Add a `transport` option to `#[AsMessageHandler]` and to the `messenger.message_handler` tag to route the handled messages to that transport and bind the handler to it
+ * Add `OutboxStamp` and `OutboxSender` to store messages in an outbox transport and forward them to their target transport when the outbox is consumed
+ * Add the `outbox` option to transports
 
 8.1
 ---